First major API rewrite (you didn't think it would be stable did you?-)
This revision introduces a separation between the sampling algorithm
and the gsl_sampler struct similar to the separation between the
gsl_rng struct and the gsl_rng_type.

The gsl_sampler uses the new gsl_sampling_records struct to store the
total and remaining numbers of records (overall, and in the sample
being constructed).  It also includes a void *state element in order
to keep track of values that must be maintained across calls to the
skip function.

(The gsl_sampling_records struct is maybe a bit wanky, and requires
extra malloc calls, but meh.  It can always be removed later and
allows for some parts of the code to be more readable.)

The sampling process must now be initialised with the gsl_sampler_init
function, specifying in advance the total number of records and the
desired sample size.

gsl_sampler_alloc and gsl_sampler_free functions have been added
for (surprise!-) allocting and freeing the new gsl_sampler structs.